requirements.txt
Django
django_rest_framework
pandas
xlrd
Pillow